first attempt to modularize my code
