+++ journal/db.py
@@ -1,25 +1,16 @@
 import sys
+from .models import JournalEntry
 import json
-from dataclasses import dataclass, asdict, field
-from datetime import datetime
-from pathlib import Path
 import typer
+from pathlib import Path
+from datetime import datetime
 from uuid import uuid4
+from dataclasses import dataclass, asdict, field
 
 
 # 📁 File where journal entries are stored
 DB_FILE = Path("journal.json")
 DB_FILE.touch(exist_ok=True)
-
-
-# 📝 Dataclass representing a journal entry
-@dataclass
-class JournalEntry:
-    title: str
-    content: str
-    date: str = field(default_factory=lambda: datetime.now().isoformat())
-    tags_list: list[str] = field(default_factory=list)
-    id: str = field(default_factory=lambda: str(uuid4()))
 
 
 # 📥 Load entries from the file
@@ -53,72 +44,12 @@ def save_entries(entries):
     with open(DB_FILE, "w", encoding="utf-8") as file:
         json.dump([asdict(entry) for entry in entries], file, indent=2)
 
-
-# type for the command line interface
-app = typer.Typer()
-
-
 # Extracted function
 def add_entry(title: str, content: str, tags_list: list) -> None:
     entries = load_entries()
     new_entry = JournalEntry(title=title, content=content, tags_list=tags_list)
     entries.append(new_entry)
     save_entries(entries)
-
-
-# ➕ Add a new entry
-@app.command()
-def add(
-    title: str = typer.Option(..., prompt="📝 Title"),
-    content: str = typer.Option(..., prompt="📓 Content"),
-    tags: str = typer.Option(
-        "", prompt="🏷️ Tags (comma-separated)", help="Optional tags for the entry    "
-    ),
-):
-    """➕ Add a new journal entry."""
-    tags_list = [tag.strip() for tag in tags.split(",") if tag.strip()] if tags else []
-    add_entry(title, content, tags_list)
-    typer.echo(f"✅ Added entry: '{title}' successfully.")
-
-
-# 📘 List journal entries
-@app.command()
-def list():
-    """📘 List all journal entries."""
-    entries = load_entries()
-    if not entries:
-        typer.echo("📭 No entries found.")
-    else:
-        typer.echo("📘 Your Journal:")
-        n = 0
-        # Iterate through entries and print them
-        for i, entry in enumerate(entries, start=1):
-            n = i
-            print(f"\n{n} 📅 {entry.date}")
-            print(f"📓 {entry.title}")
-            print(entry.content)
-            print(
-                f"🏷️ Tags: {', '.join(entry.tags_list) if entry.tags_list else 'None'}"
-            )
-            print(f"ID: {entry.id}")
-            print("-" * 30)
-
-
-# 📊 Count total entries
-@app.command()
-def count():
-    """📊 Count total journal entries."""
-    entries = load_entries()
-    count = len(entries)
-    typer.echo(f"\n 📊 Total entries 👁  in the Journal🟰 {count}")
-
-
-# Exit the application
-@app.command()
-def exit():
-    """🙌🏼 Exit the Journal application."""
-    typer.echo("🙌🏼 Thank you for using the Journal! Goodbye!")
-    typer.Exit(code=0)
 
 
 # Help function (not decorated as command, just internal)
@@ -172,24 +103,9 @@ def display_results(results):
             print("-" * 30)
 
 
-@app.command()
-def search(query: str = typer.Option(..., prompt="🔍 Enter search query")):
-    """🔍 Search for entries by query."""
-    entries = load_entries()
-    results = find_entries(entries, query)
-    display_results(results)
-
-
-@app.command()
-def query_tag(tag: str = typer.Option(..., prompt="🏷️ Enter tag to filter by")):
-    """🔍 Filter entries by tag."""
-    entries = load_entries()
-    results = find_tags(entries, tag)
-    display_results(results)
-
-
 # Interactive menu for the Journal application
 def interactive_menu():
+    from .cli import add, list, count, exit, search, query_tag
     typer.echo(
         "Welcome to your Journal! 📖"
         "\nChoose an option:"
@@ -240,15 +156,3 @@ def interactive_menu():
         typer.echo("\n👋 Exiting...")
     except Exception as e:
         typer.echo(f"⚠️ Error: {e}")
-
-
-# 🏁 Main entry point for the application
-if __name__ == "__main__":
-    # If no arguments are provided, run the interactive menu
-    if len(sys.argv) == 1:
-        interactive_menu()
-
-    else:
-        app()
-        # If arguments are provided, run the app normally
-        # This allows the app to be used in both interactive and non-interactive modes
